Changed Docker image startups to use absolute path so that working directory does not impact execution.
While unlikely, a user might set their working directory as part of their volume mount or other reason. This change allows the image to continue to work
without requiring the user to change how their images are run.

Another approach would be to let it happen and explain to a user why they need to change their command, but it might be easier to have it "just work"(tm).